+++ R/units.R
@@ -145,6 +145,12 @@ unit.ctmm <- function(CTMM,length=1,time=1)
       }
     }
 
+    if("error" %in% NAMES)
+    {
+      CTMM$COV["error",] <- CTMM$COV["error",]/length^2
+      CTMM$COV[,"error"] <- CTMM$COV[,"error"]/length^2
+    }
+
     if("circle" %in% NAMES)
     {
       CTMM$COV["circle",] <- CTMM$COV["circle",] * time